task: rename sidebar component
